requirements.txt
# numpy==1.12.1
# matplotlib==2.1.0
# pillow==5.2.0
# scikit-learn==0.19.1
# scikit-image==0.14.2
# scipy==1.2.1
# h5py==2.7.1
# eventlet==0.22.0
# flask-socketio==2.9.2
# seaborn==0.8.1
# pandas==0.23.3
# ffmpeg==3.4.1
# imageio==2.1.2
# pyqt==5.6.0
# moviepy==0.2.3.2
# opencv-python
# requests==2.18.4
# tensorflow==2.15.0
# keras==2.15.0
# numpy==1.16.0
# matplotlib==2.1.0
pillow==5.2.0
scikit-learn==0.19.1
scikit-image==0.14.2
scipy==1.2.1
h5py==2.7.1
eventlet==0.22.0
flask-socketio==2.9.2
seaborn==0.8.1
pandas==0.23.3
ffmpeg==3.4.1
imageio==2.1.2
pyqt==5.6.0
moviepy==0.2.3.2
opencv-python
requests==2.18.4
# tensorflow==2.15.0
# keras==2.15.0
